# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KHARAGOLF Enterprise - Golf Tournament SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/tournaments/{tournamentId}/players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List registered players in a tournament
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: registerPlayer
      tags: [players]
      summary: Register a player for a tournament
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPlayerInput"
      responses:
        "201":
          description: Registered player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /organizations/{orgId}/tournaments/{tournamentId}/players/{playerId}:
    put:
      operationId: updatePlayer
      tags: [players]
      summary: Update player registration
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPlayerInput"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: removePlayer
      tags: [players]
      summary: Remove player from tournament
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Player:
      type: object
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        userId:
          type: integer
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        handicapIndex:
          type: number
          nullable: true
        ghinNumber:
          type: string
          nullable: true
        flight:
          type: string
          nullable: true
        teeBox:
          type: string
          nullable: true
        paymentStatus:
          type: string
          enum: [unpaid, pending, paid, refunded]
        paymentLinkUrl:
          type: string
          nullable: true
        paymentLinkId:
          type: string
          nullable: true
        razorpayPaymentId:
          type: string
          nullable: true
        dns:
          type: boolean
          nullable: true
        checkedIn:
          type: boolean
        checkedInAt:
          type: string
          nullable: true
        teamName:
          type: string
          nullable: true
        currentRound:
          type: integer
        currentHole:
          type: integer
          nullable: true
        grossScore:
          type: integer
          nullable: true
        netScore:
          type: number
          nullable: true
        scoreToPar:
          type: integer
          nullable: true
        registeredAt:
          type: string
      required: [id, tournamentId, firstName, lastName, paymentStatus, checkedIn, currentRound, registeredAt]
    RegisterPlayerInput:
      type: object
      properties:
        userId:
          type: integer
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        handicapIndex:
          type: number
          nullable: true
        ghinNumber:
          type: string
          nullable: true
        flight:
          type: string
          nullable: true
        teeBox:
          type: string
          nullable: true
        teamName:
          type: string
          nullable: true
      required: [firstName, lastName]
...